+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 * 🚀 Add `@@asyncIterator` to `ReadableStream` ([#11](https://github.com/MattiasBuelens/web-streams-polyfill/pull/11))
 * 🚀 Add `polyfill/es2018` and `ponyfill/es2018` variants ([#11](https://github.com/MattiasBuelens/web-streams-polyfill/pull/11))
+* 🐛 Fix using unsupported `Object.assign` on Internet Explorer
 * 👓 Align with [spec version `bf2cac8`](https://github.com/whatwg/streams/tree/bf2cac8a52664df3e6da7a48755890e87b00953a/) ([#11](https://github.com/MattiasBuelens/web-streams-polyfill/pull/11))
 
 ## v0.3.1 (2019-02-25)